Improve cross-request flow and CSV parsing
Link notifications to active cross-requests and add atomic advancement. Store inserted store_notifications.id and track current_request_notification_id; prefer matching responses by response_to_request_id (with new DB column and index) while keeping title fallback. Add advance_cross_request PL/pgSQL function to atomically advance a step (FOR UPDATE) and insert the outgoing notification; client uses this RPC with a DB-backed fallback to avoid duplicate notifications. Change aggregate_notification_id to BIGINT and cast ids in migrations; ensure shared_inventory updates clamp quantities >= 0. Also clamp parsed CSV quantities to a non-negative value.

+++ supabase/migrations/add_active_cross_requests.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS active_cross_requests (
   current_index INT DEFAULT 0,
   timer_started TIMESTAMPTZ NOT NULL DEFAULT now(),
   status TEXT DEFAULT 'active',
-  aggregate_notification_id INT,
+  aggregate_notification_id BIGINT,
   created_at TIMESTAMPTZ DEFAULT now()
 );
